Virtua lCounter: fix calculation for hybrid hierarchy
