+++ packages/modules/common/store/_counter.py
@@ -89,7 +89,7 @@ class PurgeCounterState:
                 self.exported += element.data.get.exported
 
             counter_all = data.data.counter_all_data
-            elements = counter_all.get_entry_of_element(self.delegate.delegate.num)["children"]
+            elements = counter_all.get_elements_for_downstream_calculation(self.delegate.delegate.num)
             for element in elements:
                 if element["type"] == ComponentType.CHARGEPOINT.value:
                     chargepoint = data.data.cp_data[f"cp{element['id']}"]
